# Only allocate text transformer once.