@TwinProduction 2020-12-20
Released in [v1.0.1](https://github.com/TwinProduction/go-away/releases/tag/v1.0.1)